autodata/requirements.txt
gcloud-aio-storage>=9.0.0
aiohttp>=3.9.0
